requirements.txt
accelerate==0.19.0
async-generator==1.10
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
filelock==3.12.0
fsspec==2023.5.0
h11==0.14.0
huggingface-hub==0.15.1
idna==3.4
joblib==1.2.0
numpy==1.24.3
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
psutil==5.9.5
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
tqdm==4.65.0
transformers==4.29.2
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.2
webdriver-manager==3.8.6
wsproto==1.2.0
